+++ lib/canonical_local_sysaix.h
@@ -1,2 +1,2 @@
-  (int)(long)&((struct stringpool_t *)0)->stringpool_str307,
-  (int)(long)&((struct stringpool_t *)0)->stringpool_str543,
+  (int)(ptrdiff_t)&((struct stringpool_t *)0)->stringpool_str307,
+  (int)(ptrdiff_t)&((struct stringpool_t *)0)->stringpool_str543,
